1.2.12.0-alpha
changes:
- added new perk - Mugger. (15% xp amplifier, 10% speed amplifier)

+++ Assets/Scripts/Player/UI/P_XPController.cs
@@ -49,7 +49,7 @@ public class P_XPController : MonoBehaviour
             }
         }
     }
-    public void TakeXP(int takenXP)
+    public void TakeXP(float takenXP)
     {
         totalXP += takenXP;
     }
